server/requirements.txt
line-bot-sdk
flask
flask_sqlalchemy
mysqlclient
pydrive
pixivpy
feedparser
imgurpython
